requirements.txt
selenium==3.141.0
pandas==1.1.5
pyyaml==6.0
lxml==4.7.1
tqdm==4.64.0